Change package name and improve the API
limooapiclient seemds like a more appropriate package name to be coupled with limoo-api-python-client as the project name.

The message creation API was improved to support sending replies and messages
in threads or both.
The API for getting a user was improved to support None as user_id which is
basically the same thing as giving 'self' or nothing at all.