Improve the message creation API
It turns out that to attach a file to a message, you only need the hash and the
name of the file uploaded to the limoo file server. Changed the message creation
method to only expect this information. This should make it possible to use the
object returned from a file upload directly for attachment.