Fix a file upload bug by changing the interface thereof
For uploading files, we expected to receive an `io.IOBase` which then got closed
automatically by `aiohttp`. This was fine when we attempted the upload only once
because we were already authenticated, but not so well otherwise: when we are
not authenticated, on our first attempt we get a HTTP 401 and the I/O stream
would be closed and since we would use the same I/O stream object for the second
attempt, we would get a `ValueError` telling us we can't use a closed stream. To
fix this problem, we now ask for the path to the file to be uploaded, and we
open a stream for it every time we want to attempt an upload. While this fixes
the problem, it also means that we can no longer accept an arbitrary `io.IOBase`
object; but for now, that should be okay.