Fix the bugs manifesting when working with a server designated by an IP address
By default, `aiohttp` does not accept cookies from servers designated by an IP
address [1]. So if we wanted to connect to a Limoo server by its IP address, our
authentication would be futile and all API requests would get a
`LimooAuthenticationFailed`. According to `aiohttp`, doing this is required by
RFC2109 which is already obsoleted RFC2965 (which is futher obsoleted by
RFC6265) but also, I didn't find any such a requirement anywhere in that RFC. We
now use an unsafe version of `aiohttp.CookieJar` to avoid this default behavior.

[1] https://docs.aiohttp.org/en/stable/client_advanced.html#cookie-safety